+++ tartufo/config.py
@@ -241,7 +241,7 @@ def compile_path_rules(patterns: Iterable[str]) -> List[Pattern]:
     ]
 
 
-def compile_rules(patterns: Iterable[Dict[str, str]]) -> List[Rule]:
+def compile_rules(patterns: Iterable[Dict[str, str]], exclude_type: str) -> List[Rule]:
     """Take a list of regex string with paths and compile them into a List of Rule.
 
     :param patterns: The list of patterns to be compiled
@@ -255,12 +255,17 @@ def compile_rules(patterns: Iterable[Dict[str, str]]) -> List[Rule]:
             raise ConfigException(
                 f"Invalid value for match-type: {pattern.get('match-type')}"
             ) from exc
-        try:
-            scope = Scope(pattern.get("scope", Scope.Line.value))
-        except ValueError as exc:
-            raise ConfigException(
-                f"Invalid value for scope: {pattern.get('scope')}"
-            ) from exc
+        if exclude_type == "regex":
+            # regex exclusions always have line scope
+            scope = Scope.Line
+        else:
+            # entropy exclusions can specify scope
+            try:
+                scope = Scope(pattern.get("scope", Scope.Line.value))
+            except ValueError as exc:
+                raise ConfigException(
+                    f"Invalid value for scope: {pattern.get('scope')}"
+                ) from exc
         try:
             rules.append(
                 Rule(
@@ -273,6 +278,6 @@ def compile_rules(patterns: Iterable[Dict[str, str]]) -> List[Rule]:
             )
         except KeyError as exc:
             raise ConfigException(
-                f"Invalid exclude-entropy-patterns: {patterns}"
+                f"Invalid exclude-{exclude_type}-patterns: {patterns}"
             ) from exc
     return rules